Move bootstrap test to airframe-config
